\name{dimension_province}
\alias{dimension_province}
\docType{data}
\title{The province dimension or lookup table}
\description{This table lists all provinces}
\usage{data("dimension_province")}
\format{
  A data frame with 9 observations on the following 2 variables.
  \itemize{
    \item province_key. a unique province key
    \item province_name. the common name of the province
  }
}
\details{None}
\source{}
\references{}
\examples{
data(dimension_province)
## maybe str(dimension_province) ; plot(dimension_province) ...
}
\keyword{datasets}
